requirements.txt
# Requirements for the project
beautifulsoup4==4.14.3
Django==5.2.11
django-taggit==6.1.0
gmpy2
lief==0.17.3
olefile==0.47
oletools==0.60.2
opencv-python==4.13.0.92
pandas==3.0.1
pdfminer
pdfminer.six==20260107
pefile==2024.8.26
prettytable==3.17.0
pillow==12.1.1
pycryptodome==3.23.0
PyExifTool==0.5.6
PyMuPDF==1.27.1
pypdf>=3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzipper==0.3.6
py7zr==1.1.0
requests==2.32.5
shodan==1.31.0
tabulate==0.9.0
tldextract==5.3.1
yara-python==4.5.4
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-cors-headers==4.7.0
gunicorn>=22.0.0
whitenoise[brotli]>=6.8.0
psycopg2-binary>=2.9.10
capstone>=5.0.3
